Compile as x86
ILC (the Native AOT compiler) is a crosscompiler and ships with x86 support even though we don't actually have it fully implemented for Native AOT in mainstream .NET. Since SmolSharp only uses the codegen (RyuJIT) and RyuJIT supports x86 just fine, we can still generate x86.

This has some hackery though:

* We can't publish for the win-x64 RID because those packages are missing. So keep publishing as win-x64, but whack it to x86 when NativeAOT logic becomes concerned.
* x86 requires a bunch of assembly helpers with non-standard calling conventions. We need to write them in assembly :(
* x86 has a special mangling for x86 that ILC doesn't implement (I tried to implement it in the past in https://github.com/dotnet/corert/pull/7943). We need to whack the mangling into imported names. To find the right mangled name, just search strings in the appropriate LIB file (e.g. kernel32.lib).

I'm submitting it as a draft. It is kind of hacky. No hard feelings if you don't want to merge it.

The size numbers are much better on x86: 1711 bytes for HelloWorld, 2299 bytes for Mandelbrot, 5832 bytes for Ocean.